Backend/requirements.txt
Flask
psycopg2-binary
python-dotenv
flask-cors
scikit-learn
joblib
numpy
pandas
requests



# APIs
OpenWeatherMap API



# (Image succesfully uplode command) {run in terminal aftar start server, we can't use postman} =
# curl.exe -X POST -F "image=@C:\Users\divya\Desktop\SmartKrishi\Backend\uploads\plants\00000007.jpg" http://127.0.0.1:5000/upload-image
